updated java okhttp-gson-parcelable-model sample.

+++ samples/client/petstore/java/okhttp-gson-parcelableModel/src/test/java/io/swagger/client/api/PetApiTest.java
@@ -66,7 +66,7 @@ public class PetApiTest {
     @Test
     public void deletePetTest() throws ApiException {
         
-        Integer petId = null;
+        Long petId = null;
         
         String apiKey = null;
         
@@ -121,8 +121,8 @@ public class PetApiTest {
      */
     @Test
     public void getPetByIdTest() throws ApiException {
-        
-        Integer petId = null;
+
+        Long petId = null;
         
         Pet response = api.getPetById(petId);
 
@@ -157,8 +157,8 @@ public class PetApiTest {
      */
     @Test
     public void updatePetWithFormTest() throws ApiException {
-        
-        Integer petId = null;
+
+        Long petId = null;
         
         String name = null;
 
@@ -179,8 +179,8 @@ public class PetApiTest {
      */
     @Test
     public void uploadFileTest() throws ApiException {
-        
-        Integer petId = null;
+
+        Long petId = null;
         
         String additionalMetadata = null;
